% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-helpers.R
\name{format_query_param}
\alias{format_query_param}
\title{Format Query Parameters}
\usage{
format_query_param(param, value)
}
\arguments{
\item{param}{The parameter name. Can be a character string, e.g. "userID", or
a character vector, such as c("userID", "favorite_food").}

\item{value}{The parameter value. Can be a character string, "jdtrat", or a
character vector, such as c("jdtrat", "sushi").}
}
\value{
A formatted query string
}
\description{
This is a utility function used in \code{\link{insert_query_param}}. Given a
single query parameter-value pair, or ordered character vectors, it will
properly format and return a query string.
}
\examples{

# Returns proper formatting for single query parameters: "?person=jdtrat"
format_query_param(param = "person",
                   value = "jdtrat")

# Returns proper formatting for multiple query parameters:
# "?person=jdtrat&favorite_food=sushi&best_cat=tucker"
format_query_param(param = c("person", "favorite_food", "best_cat"),
                   value = c("jdtrat", "sushi", "tucker")
                   )

}
\seealso{
Other Query Strings: 
\code{\link{extract_query_param}()},
\code{\link{insert_query_param}()}
}
\concept{Query Strings}
