% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeButton.R
\name{random_link}
\alias{random_link}
\title{Get a Random Link For Studies with Multiple Treatment Groups}
\usage{
random_link(groupLinkPairs, addQueryParameter = TRUE)
}
\arguments{
\item{groupLinkPairs}{A data frame with two columns "group" and "link"
containing the group and URL links for randomization. See examples for more
details.}

\item{addQueryParameter}{LOGICAL: TRUE and a query parameter "group_id" will
be added to the randomization URL. FALSE and it won't.}
}
\value{
A link based on key-value pairs for treatment groups.
}
\description{
This function takes in a data frame of key-value paris for treatment groups
and links and returns a link that can be passed into the "onclick" attribute
of HTML buttons (e.g. \code{\link[shiny:actionButton]{shiny::actionButton()}}).
}
\examples{

keys <- data.frame(
  group = c("A", "B"),
  link = c("https://yourwebsite.com/this-task",
  "https://yourwebsite.com/this-other-task")
)

if (interactive()) {

 link <- random_link(groupLinkPairs = keys,
 addQueryParameter = TRUE)

 shiny::actionButton(
   inputId = "randomId",
   label = "Click me!",
   onclick = paste0("window.location.replace('", link, "')"),
 )

}

}
\seealso{
randomizeButton
}
