% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeButton.R
\name{randomizeButton}
\alias{randomizeButton}
\title{Action Button that Sends Users to Different Locations}
\usage{
randomizeButton(inputId, label, groupLinkPairs, addQueryParameter = TRUE, ...)
}
\arguments{
\item{inputId}{Shiny input ID}

\item{label}{The button label}

\item{groupLinkPairs}{A data frame with two columns "group" and "link"
containing the group and URL links for randomization. See examples for more
details.}

\item{addQueryParameter}{LOGICAL: TRUE and a query parameter "group_id" will
be added to the randomization URL. FALSE and it won't.}

\item{...}{Additional parameters to pass into \code{\link[shiny:actionButton]{shiny::actionButton()}}}
}
\value{
The UI for an action button.
}
\description{
Randomization is common to behavioral data collection. This action button
allows you to supply a key-value pair of group and URL links for randomizing
participants in a {shiny} application.
}
\examples{

keys <- data.frame(
  group = c("A", "B"),
  link = c("https://yourwebsite.com/this-task",
  "https://yourwebsite.com/this-other-task")
)


if (interactive()) {

 library(shiny)

 ui <- fluidPage(
   randomizeButton(inputId = "myId",
                   label = "Click me to randomize",
                   groupLinkPairs = keys,
                   addQueryParameter = TRUE)
 )

 server <- function(input, output, session) {

 }

 shinyApp(ui, server)


}


}
